refactor(ui): moves sidebar to it's own component
